change Qt.Checked to Qt.CheckState.Checked.value
For some reason the installed version of PySide6 I built this on was 6.3.2, which makes a huge difference when the latest is 6.9.1. Apparently there is some difference in the CheckState for comparison - some problem with comparing the "state" integer to the enum of Qt.Checked. So to avoid further issue we're comparing the state integer to Qt.CheckState.Checked.value which is an equal integer.

Due for testing on Linux